Refactor creating paths to be more flexible and resilient
The IsApplicable and Create methods get the full path and check if there is something at the beginning
of the path they apply to. The Create method outputs a new path, which is the old path minus the part
that was used to create the PathElement.
This removes code duplication, because EnumerableAccess now does not have to access the property,
because Property already does this.
It also makes it more resilient and flexible. Now a path starting with "[0]" is valid, since EnumerableAccess
would simply pick it up and access the target with it.